+++ packages/cli/src/utils/mod.rs
@@ -1,4 +1,2 @@
 pub mod files;
-pub mod github;
 pub mod listing;
-pub mod net;